clasificador_intenciones/requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
python-multipart
pypdf 
langchain-huggingface
chromadb
sentence-transformers
